delete med commit

+++ backend/src/controllers/medicineController.js
@@ -56,15 +56,23 @@ const updateMedicine = async (req, res) => {
 }
 
 const deleteMedicine = async (req, res) => {
-    try {
-        const deletedMedicine = await Medicine.findByIdAndDelete(req.params.id);
-        if (!deletedMedicine) {
-          return res.status(404).json({ error: 'Medicine not found' });
-        }
-        res.status(200).json(deletedMedicine);
-      } catch (error) {
-        res.status(500).json({ error: error.message });
-    }
+  const medicine = await Medicine.findById(req.params.id)
+
+  if (!medicine) {
+    return res.status(404).json({ error: 'Medicine not found'})
+  }
+  if (req.user.id !== medicine.user.toString()) {
+    return res.status(401).json({ error: 'You can only delete your own medicines'})
+  }
+  // console.log('req.user.id:', req.user.id);
+  // console.log('medicine.user:', medicine.user);
+  try {
+    await Medicine.findByIdAndDelete(req.params.id)
+    res.status(200).json({ message: 'Medicine deleted successfully'});
+  } catch (error) {
+    res.status(500).json({ error: error.message });
+  }
 }
+
 
 module.exports = { addMedicine, getMedicines, getMedicine, updateMedicine, deleteMedicine }
